requirements.txt
# Flask and extensions
Flask==3.0.0
Werkzeug==3.0.1

# Database
supabase==2.9.0
postgrest==0.17.1

# Authentication & Security
bcrypt==4.1.2

# File handling
Pillow==10.2.0

# Email
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# HTTP requests
requests==2.31.0

# WSGI server for production
gunicorn==21.2.0

# PDF generation
reportlab==4.0.9

# Excel generation
openpyxl==3.1.2
